+++ packages/components/src/components.d.ts
@@ -7,9 +7,10 @@
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { Provider, StyleSettingsMap, TimeQuery, TimeSeriesData, TimeSeriesDataRequest, TimeSeriesDataRequestSettings, TreeQuery, Viewport } from "@iot-app-kit/core";
 import { AlarmsConfig, Annotations, Axis, LabelsConfig, LayoutConfig, LegendConfig, MessageOverrides, MinimalSizeConfig, MovementConfig, ScaleConfig, TableColumn, Trend } from "@synchro-charts/core";
+import { Item, RecursivePartial, TableItem, TableMessages, TableProps } from "@iot-app-kit/table";
 import { BranchReference, SiteWiseAssetTreeNode } from "@iot-app-kit/source-iotsitewise";
 import { ColumnDefinition, FilterTexts } from "./components/iot-resource-explorer/types";
-import { TableProps } from "@awsui/components-react/table";
+import { TableProps as TableProps1 } from "@awsui/components-react/table";
 import { EmptyStateProps, ITreeNode, UseTreeCollection } from "@iot-app-kit/related-table";
 import { NonCancelableCustomEvent } from "@awsui/components-react";
 export namespace Components {
@@ -60,6 +61,13 @@ export namespace Components {
         "viewport": Viewport;
         "widgetId": string;
     }
+    interface IotReactTable {
+        "columnDefinitions": TableProps['columnDefinitions'];
+        "items": TableItem[];
+        "messageOverrides": TableMessages;
+        "propertyFiltering": TableProps['propertyFiltering'];
+        "sorting": TableProps['sorting'];
+    }
     interface IotResourceExplorer {
         "columnDefinitions": ColumnDefinition<any>[];
         "empty"?: EmptyStateProps;
@@ -67,10 +75,10 @@ export namespace Components {
         "filterEnabled": boolean;
         "filterTexts"?: FilterTexts;
         "loadingText"?: string;
-        "onSelectionChange": (event: NonCancelableCustomEvent<TableProps.SelectionChangeDetail<unknown>>) => void;
+        "onSelectionChange": (event: NonCancelableCustomEvent<TableProps1.SelectionChangeDetail<unknown>>) => void;
         "paginationEnabled": boolean;
         "query": TreeQuery<SiteWiseAssetTreeNode[], BranchReference>;
-        "selectionType"?: TableProps.SelectionType;
+        "selectionType"?: TableProps1.SelectionType;
         "sortingEnabled": boolean;
         "widgetId": string;
         "wrapLines": boolean;
@@ -125,9 +133,13 @@ export namespace Components {
     }
     interface IotTable {
         "annotations": Annotations;
-        "messageOverrides"?: MessageOverrides;
+        "columnDefinitions": TableProps['columnDefinitions'];
+        "items": Item[];
+        "messageOverrides"?: RecursivePartial<TableMessages>;
+        "propertyFiltering": TableProps['propertyFiltering'];
         "queries": TimeQuery<TimeSeriesData[], TimeSeriesDataRequest>[];
         "settings": TimeSeriesDataRequestSettings;
+        "sorting": TableProps['sorting'];
         "styleSettings": StyleSettingsMap | undefined;
         "tableColumns": TableColumn[];
         "trends": Trend[];
@@ -144,9 +156,9 @@ export namespace Components {
         "styleSettings": StyleSettingsMap | undefined;
     }
     interface IotTreeTable {
-        "ariaLabels": TableProps.AriaLabels<unknown>;
+        "ariaLabels": TableProps1.AriaLabels<unknown>;
         "collectionOptions": UseTreeCollection<unknown>;
-        "columnDefinitions": TableProps.ColumnDefinition<any>[];
+        "columnDefinitions": TableProps1.ColumnDefinition<any>[];
         "empty": EmptyStateProps;
         "expanded": boolean;
         "filterPlaceholder": string;
@@ -155,10 +167,10 @@ export namespace Components {
         "loading": boolean;
         "loadingText": string;
         "onExpandChildren": (node: ITreeNode<any>) => void;
-        "onSelectionChange": (event: NonCancelableCustomEvent<TableProps.SelectionChangeDetail<unknown>>) => void;
-        "onSortingChange": (event: NonCancelableCustomEvent<TableProps.SortingState<unknown>>) => void;
+        "onSelectionChange": (event: NonCancelableCustomEvent<TableProps1.SelectionChangeDetail<unknown>>) => void;
+        "onSortingChange": (event: NonCancelableCustomEvent<TableProps1.SortingState<unknown>>) => void;
         "resizableColumns": boolean;
-        "selectionType": TableProps.SelectionType;
+        "selectionType": TableProps1.SelectionType;
         "sortingDisabled": boolean;
         "wrapLines": boolean;
     }
@@ -187,6 +199,12 @@ declare global {
     var HTMLIotLineChartElement: {
         prototype: HTMLIotLineChartElement;
         new (): HTMLIotLineChartElement;
+    };
+    interface HTMLIotReactTableElement extends Components.IotReactTable, HTMLStencilElement {
+    }
+    var HTMLIotReactTableElement: {
+        prototype: HTMLIotReactTableElement;
+        new (): HTMLIotReactTableElement;
     };
     interface HTMLIotResourceExplorerElement extends Components.IotResourceExplorer, HTMLStencilElement {
     }
@@ -264,6 +282,7 @@ declare global {
         "iot-bar-chart": HTMLIotBarChartElement;
         "iot-kpi": HTMLIotKpiElement;
         "iot-line-chart": HTMLIotLineChartElement;
+        "iot-react-table": HTMLIotReactTableElement;
         "iot-resource-explorer": HTMLIotResourceExplorerElement;
         "iot-resource-explorer-demo": HTMLIotResourceExplorerDemoElement;
         "iot-scatter-chart": HTMLIotScatterChartElement;
@@ -326,6 +345,13 @@ declare namespace LocalJSX {
         "viewport": Viewport;
         "widgetId"?: string;
     }
+    interface IotReactTable {
+        "columnDefinitions": TableProps['columnDefinitions'];
+        "items": TableItem[];
+        "messageOverrides"?: TableMessages;
+        "propertyFiltering"?: TableProps['propertyFiltering'];
+        "sorting"?: TableProps['sorting'];
+    }
     interface IotResourceExplorer {
         "columnDefinitions"?: ColumnDefinition<any>[];
         "empty"?: EmptyStateProps;
@@ -333,10 +359,10 @@ declare namespace LocalJSX {
         "filterEnabled"?: boolean;
         "filterTexts"?: FilterTexts;
         "loadingText"?: string;
-        "onSelectionChange"?: (event: NonCancelableCustomEvent<TableProps.SelectionChangeDetail<unknown>>) => void;
+        "onSelectionChange"?: (event: NonCancelableCustomEvent<TableProps1.SelectionChangeDetail<unknown>>) => void;
         "paginationEnabled"?: boolean;
         "query"?: TreeQuery<SiteWiseAssetTreeNode[], BranchReference>;
-        "selectionType"?: TableProps.SelectionType;
+        "selectionType"?: TableProps1.SelectionType;
         "sortingEnabled"?: boolean;
         "widgetId"?: string;
         "wrapLines"?: boolean;
@@ -391,9 +417,13 @@ declare namespace LocalJSX {
     }
     interface IotTable {
         "annotations"?: Annotations;
-        "messageOverrides"?: MessageOverrides;
+        "columnDefinitions": TableProps['columnDefinitions'];
+        "items": Item[];
+        "messageOverrides"?: RecursivePartial<TableMessages>;
+        "propertyFiltering"?: TableProps['propertyFiltering'];
         "queries": TimeQuery<TimeSeriesData[], TimeSeriesDataRequest>[];
         "settings"?: TimeSeriesDataRequestSettings;
+        "sorting"?: TableProps['sorting'];
         "styleSettings"?: StyleSettingsMap | undefined;
         "tableColumns"?: TableColumn[];
         "trends"?: Trend[];
@@ -410,9 +440,9 @@ declare namespace LocalJSX {
         "styleSettings"?: StyleSettingsMap | undefined;
     }
     interface IotTreeTable {
-        "ariaLabels"?: TableProps.AriaLabels<unknown>;
+        "ariaLabels"?: TableProps1.AriaLabels<unknown>;
         "collectionOptions": UseTreeCollection<unknown>;
-        "columnDefinitions": TableProps.ColumnDefinition<any>[];
+        "columnDefinitions": TableProps1.ColumnDefinition<any>[];
         "empty"?: EmptyStateProps;
         "expanded"?: boolean;
         "filterPlaceholder"?: string;
@@ -421,10 +451,10 @@ declare namespace LocalJSX {
         "loading"?: boolean;
         "loadingText"?: string;
         "onExpandChildren"?: (node: ITreeNode<any>) => void;
-        "onSelectionChange"?: (event: NonCancelableCustomEvent<TableProps.SelectionChangeDetail<unknown>>) => void;
-        "onSortingChange"?: (event: NonCancelableCustomEvent<TableProps.SortingState<unknown>>) => void;
+        "onSelectionChange"?: (event: NonCancelableCustomEvent<TableProps1.SelectionChangeDetail<unknown>>) => void;
+        "onSortingChange"?: (event: NonCancelableCustomEvent<TableProps1.SortingState<unknown>>) => void;
         "resizableColumns"?: boolean;
-        "selectionType"?: TableProps.SelectionType;
+        "selectionType"?: TableProps1.SelectionType;
         "sortingDisabled"?: boolean;
         "wrapLines"?: boolean;
     }
@@ -438,6 +468,7 @@ declare namespace LocalJSX {
         "iot-bar-chart": IotBarChart;
         "iot-kpi": IotKpi;
         "iot-line-chart": IotLineChart;
+        "iot-react-table": IotReactTable;
         "iot-resource-explorer": IotResourceExplorer;
         "iot-resource-explorer-demo": IotResourceExplorerDemo;
         "iot-scatter-chart": IotScatterChart;
@@ -459,6 +490,7 @@ declare module "@stencil/core" {
             "iot-bar-chart": LocalJSX.IotBarChart & JSXBase.HTMLAttributes<HTMLIotBarChartElement>;
             "iot-kpi": LocalJSX.IotKpi & JSXBase.HTMLAttributes<HTMLIotKpiElement>;
             "iot-line-chart": LocalJSX.IotLineChart & JSXBase.HTMLAttributes<HTMLIotLineChartElement>;
+            "iot-react-table": LocalJSX.IotReactTable & JSXBase.HTMLAttributes<HTMLIotReactTableElement>;
             "iot-resource-explorer": LocalJSX.IotResourceExplorer & JSXBase.HTMLAttributes<HTMLIotResourceExplorerElement>;
             "iot-resource-explorer-demo": LocalJSX.IotResourceExplorerDemo & JSXBase.HTMLAttributes<HTMLIotResourceExplorerDemoElement>;
             "iot-scatter-chart": LocalJSX.IotScatterChart & JSXBase.HTMLAttributes<HTMLIotScatterChartElement>;